i2c: cadence: Refactor internal structs into a private header
To improve code organization and prepare for a subsequent patch that
adds new functionality, separate the internal data structure definitions
from the implementation logic.

Move the `struct cdns_i2c` `enum cdns_i2c_mode`
and `enum cdns_i2c_slave_state` from i2c-cadence.c into
a new private header, i2c-cadence.h.

This is purely a code movement and introduces no functional changes.
